+++ src/Method/SendRichMessage.php
@@ -1,0 +1,76 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phptg\BotApi\Method;
+
+use Phptg\BotApi\MethodInterface;
+use Phptg\BotApi\ParseResult\ValueProcessor\ObjectValue;
+use Phptg\BotApi\Transport\HttpMethod;
+use Phptg\BotApi\Type\ForceReply;
+use Phptg\BotApi\Type\InlineKeyboardMarkup;
+use Phptg\BotApi\Type\InputRichMessage;
+use Phptg\BotApi\Type\Message;
+use Phptg\BotApi\Type\ReplyKeyboardMarkup;
+use Phptg\BotApi\Type\ReplyKeyboardRemove;
+use Phptg\BotApi\Type\ReplyParameters;
+use Phptg\BotApi\Type\SuggestedPostParameters;
+
+/**
+ * @see https://core.telegram.org/bots/api#sendrichmessage
+ *
+ * @template-implements MethodInterface<Message>
+ */
+final readonly class SendRichMessage implements MethodInterface
+{
+    public function __construct(
+        private int|string $chatId,
+        private InputRichMessage $richMessage,
+        private ?string $businessConnectionId = null,
+        private ?int $messageThreadId = null,
+        private ?int $directMessagesTopicId = null,
+        private ?bool $disableNotification = null,
+        private ?bool $protectContent = null,
+        private ?bool $allowPaidBroadcast = null,
+        private ?string $messageEffectId = null,
+        private ?SuggestedPostParameters $suggestedPostParameters = null,
+        private ?ReplyParameters $replyParameters = null,
+        private InlineKeyboardMarkup|ReplyKeyboardMarkup|ReplyKeyboardRemove|ForceReply|null $replyMarkup = null,
+    ) {}
+
+    public function getHttpMethod(): HttpMethod
+    {
+        return HttpMethod::POST;
+    }
+
+    public function getApiMethod(): string
+    {
+        return 'sendRichMessage';
+    }
+
+    public function getData(): array
+    {
+        return array_filter(
+            [
+                'business_connection_id' => $this->businessConnectionId,
+                'chat_id' => $this->chatId,
+                'message_thread_id' => $this->messageThreadId,
+                'direct_messages_topic_id' => $this->directMessagesTopicId,
+                'rich_message' => $this->richMessage->toRequestArray(),
+                'disable_notification' => $this->disableNotification,
+                'protect_content' => $this->protectContent,
+                'allow_paid_broadcast' => $this->allowPaidBroadcast,
+                'message_effect_id' => $this->messageEffectId,
+                'suggested_post_parameters' => $this->suggestedPostParameters?->toRequestArray(),
+                'reply_parameters' => $this->replyParameters?->toRequestArray(),
+                'reply_markup' => $this->replyMarkup?->toRequestArray(),
+            ],
+            static fn(mixed $value): bool => $value !== null,
+        );
+    }
+
+    public function getResultType(): ObjectValue
+    {
+        return new ObjectValue(Message::class);
+    }
+}